packages/socd/src/socd/pyproject.toml
[project]
name = "socd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cuda-python[all]>=12.9.0",
    "cupy-cuda12x>=13.4.1",
]
